requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
tensorflow>=2.15.0
numpy>=1.24.0
pillow>=10.0.0
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
plotly>=5.15.0
dash>=2.14.0
dash-bootstrap-components>=1.4.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
